pyproject.toml
[tool.poetry]
name = "lab-riddle-server"
version = "0.1.0"
description = ""
authors = ["emma <eqbrown@isf.edu.hk>"]
readme = "README.md"
packages = [{include = "lab_riddle_server"}]

[tool.poetry.dependencies]
python = "^3.10"
django-banjo = "^0.2.9"
fuzzywuzzy = "^0.18.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
